env: setup export to config directory
The prior implementation using a .env file has a few issues:

* The format does not provide a great way to visualize path variables.
  It is not helpful to list the long values on single lines, and the
  introduction of line breaks appears to propagate to variable value
  itself.
* There was no determination made between user variables and machine
  variables.
* I have not found a good mechanism to consume the file in Windows.

The new approach is to pull the environment variables from the registry
(so that variables in the values aren't expanded), filter out the
built-in variables, and export to a named YAML file. The format was
chosen since it is very readable for array values and has good
PowerShell support.