env: use expanded paths on machine
After updating environment variable paths to use un-expanded variants
(e.g. "%PROGRAMFILES%\Git\cmd"), I found that installers and programs
did not recognize the variables, and would re-add or modify them.

The variables have been converted back to the expanded format and the
script has been updated to convert them back for specified prefixes. A
side benefit is that the %SYSTEMROOT% variable that was added to the
user path can be removed, since its only purpose was to trigger the
Windows dialog to recognize the text as an array.

Link: https://superuser.com/a/1594989