env: change exports into providers
The export naming limits the functionality to generation of the tracked
files. By naming the sources as providers, it allows for additional
functionality to be added without confusion.

The file used as the data store for each provider is added to the
configuration, and each export script is updated to require a target
path to be passed in. This centralizes to paths to a single
configuration file.