bin: add script to resolve a relative path
Determines the number of shared ancestor segments between two paths and
uses the count to generate a relative path from the source directory to
the target path.

The Resolve-Path cmdlet provides this functionality, but only works if
both the source and target paths exist. For the planned usage to
determine symlink targets, only the source path must exist.