env: remove env provider from machine scripts
Since environment variables are not versioned like the other providers
and the output from the export script cannot be transformed into a CSV,
this process must be independent.

It will likely still be performed after a machine export, but via a
separate call. The 'eev' PowerShell alias has been added to allow for an
easier execution in the meantime.