software: define export process
Adds the new top-level software export script, which handles the
aggregation of the individual provider exports.

Rather than saving the results in individual locations, the
determination for installation of global exports is represented by a
single CSV file, and the provider-specific installation definitions are
represented by separate YAML files.

For locally scoped exports, the CSV file is untouched, and the
installation and definition representations are combined into a
machine-specific YAML file.