git: expand log sub-alias commands
A couple log aliases used the central "l" alias as the base with
additional modifiers, but this prevents posh-git from completing branch
names since it does not perform the double-hop to determine the
underlying command (e.g. "ls" -> "l" -> "log"). The "l" alias usage in
both has been replaced with the content of the alias to allow for tab
completion.