pwsh: use inspect for file type check for indents
The prior usage of Git restricts the script to be run only on Git
repositories, and was also scoped to the dotfiles repository alone.

The inspect binary from the content_inspector crate provides a fast and
accurate (as possible) check to determine whether a binary file should
be excluded from the indentation check.